feat/follower_following make the page structure in the correct way
